- ActivityApiControllerTests no longer inherits from TestBase - deleting all helper methods and constructor  in ActivityApiControllerTests as no longer is GetActivityApiController method being used to build a real DbContext via the ServiceProvider in the in unit tests - tests that were in this class that ran against IAllReadyDataAccess.DeleteActivityAndTaskSignupsAsync using the DbContext via the ServiceProvider have been moved to the AllReadyDataAccessEF7Tests class - moving ControllerTestHelpers, AttributeExtensions, EnumerableExtensions and JsonResultExtension into Extensions folder under AllReady.UnitTest project
