+++ AllReadyApp/Web-App/AllReady.UnitTest/Features/Campaigns/CampaignQueryHandlerTests.cs
@@ -1,9 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using AllReady.Extensions;
 using AllReady.Features.Campaigns;
 using AllReady.Models;
+using AllReady.UnitTest.Extensions;
 using Moq;
 using Xunit;
